chore(home-v2): Moving varaibles out of components
