+++ src/views/Home/components/Hero.tsx
@@ -74,21 +74,22 @@ const StarsWrapper = styled.div`
   }
 `
 
+const imagePath = '/images/home/lunar-bunny/'
+const imageSrc = 'bunny'
+
+const starsImage: CompositeImageProps = {
+  path: '/images/home/lunar-bunny/',
+  attributes: [
+    { src: 'star-l', alt: '3D Star' },
+    { src: 'star-r', alt: '3D Star' },
+    { src: 'star-top-r', alt: '3D Star' },
+  ],
+}
+
 const Hero = () => {
   const { t } = useTranslation()
   const { account } = useWeb3React()
   const { theme } = useTheme()
-  const imagePath = '/images/home/lunar-bunny/'
-  const imageSrc = 'bunny'
-
-  const starsImage: CompositeImageProps = {
-    path: '/images/home/lunar-bunny/',
-    attributes: [
-      { src: 'star-l', alt: '3D Star' },
-      { src: 'star-r', alt: '3D Star' },
-      { src: 'star-top-r', alt: '3D Star' },
-    ],
-  }
 
   return (
     <>
